+++ Changes.md
@@ -1,3 +1,16 @@
+# 0.3 ( 2018-03-27 )
+
+- Added retention policies to log tables
+  Retention time is set on enable\_change\_logging using 6th argument, or you
+  can set it directly in \_flux\_tables by updating _retention_ column.
+
+- Added retention cleanup to _cleanup()_ function, so it now serves two
+  purposes:
+
+    - remove obsolete log tables
+
+    - remove old log data (rows) from log tables
+
 # 0.2 ( 2018-03-27 )
 
 - Added \_flux\_tables to extension, so on DROP EXTENSION flux it's metadata